+++ CHANGELOG.md
@@ -27,7 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-nil.
+- Add support for a line2 address field [#173](https://github.com/Shopify/worldwide/pull/173)
 
 ---